spring microservice project

+++ ticket-service/src/main/java/com/pione/ticketservice/Ticket.java
@@ -20,7 +20,7 @@ public class Ticket {
     @Id
     @GeneratedValue
     private Integer id;
-    private long eventId;
+    private Integer eventId;
     private float price;
     private boolean status;
     private String type;
